Switched storage from SQL Server to RavenDB
In order to make future changes easier the storage has been switched
from SQL Server to RavenDB. This allows for easier data-access as well
as easy searching due to the included Lucene engine.

RavenDB will, by default, run in an embedded mode. This ensures it is no
longer necessary to maintain the database and software seperately.