Feature dependencies and UML support (#69)
Two new features we're added in this release:

Navigate to dependent features from a particular feature. This is an alpha feature, which you can enable from the settings page.
Embed UML diagrams based on [nomnoml](http://www.nomnoml.com) to your feature description as an UML code block and they will be displayed in Augurk.