Add Dockerfile and associated build pipeline (#93)
A Dockerfile was added so that we can build Docker images for Augurk so that people can run Augurk very easily. Additionally the build pipeline was modified to not only produce a zip that can be used to run Augurk, but also build the Docker image and publish it to Docker Hub.

Note that the build will always publish the image straight to Docker Hub, without going through a release pipeline. It will be tagged with the major.minor.patch based on GitVersion and a -preview postfix unless it is a build on the master branch. Once we have a stable 3.0 version we should probably make clear to users that the -preview tag is basically the latest greatest.