Augurk 3.0 (#98)
This PR combines all of the work we've done for Augurk 3.0. Big feature is cross-platform support, by retargetting Augurk to run on .NET Core. This allows for running Augurk on Windows, Mac and Linux. Additionally we now ship a Docker image so getting started with Augurk is easier than ever.

Finally we've made some smaller quality of life improvements, such as:

- Readable scenario outlines by displaying the example inline when clicking on it
- Retention policy for features can be configured and will be applied retroactively
- We've fixed a bug with identical features in multiple groups where only a single feature would be shown
- Improved test coverage and various other build quality improvements
- Look forward to our new revamped website with guidance on installing Augurk 3.0 as well as migrating from earlier versions.

+semver: major