Augurk 3.1.0-Preview-1 (#221)
It has taken quite some time, but it is finally here! Augurk 3.1.0 Preview-1!
Preview? Yes, we have decided that due to our irregular schedules, from now on we will release parts of new features as soon as they are available and can be used with a reasonable stability.

For now, Augurk will run in a hybrid mode, where you can use both the original (AngularJS) UI, as well as the new features in our up-and-coming (WebAssembly) UI! While this is far from ideal, we believe it is the best approach to give you access to our newest features as soon as possible, without the need to build them in the original UI as well.

Changes in this preview:
- *Reduced storage*   We reduced the amount of storage required for your features by combining unchanged versions, without losing any actual data.
- *Preview button*   A preview button has been added to the top right of the original UI, leading you to the preview functionality.
- *Text search*   From the preview UI, it is now possible to perform a full text search on all your features. At this moment the search results only contain features which contain a complete match for the provided word. As of yet, search for partial words is not yet supported.   For now, the results will link to the features in the original UI.

Naturally, changing the data structure also broke some features. While we fully intend to fix this before going stable with 3.1.0, these features are not available in Preview-1
- *Data retention*   Preview-1 will not apply data retention and will also not retain the upload date of newly uploaded futures for future retention. We will think of a fitting solution for the features which were uploaded using this version in the stable version
- *Feature dependencies*   Feature dependencies other than the @parentFeature-tag will not display properly, as the current implementation is version-bound. We will have to redesign the storage for this feature from scratch
- *Test results*   While test results can still be uploaded and linked to your scenario’s, they will now be linked to all identical version of the tested feature instead of only to the provided version.

So, quite some changes, and we are far from done. We certainly hope to provide more functionality in the new UI, as well as fix the broken features listed above.
Until that time, we are happy to receive your feedback regarding this preview. Do not hesitate to share any feedback, be it a new thought you have or issues you encounter, through our [issue tracker](https://github.com/Augurk/Augurk/issues  ). Until the next version!

Mark & Jonathan